yet another code refactor
doing this to try and separate the responsibilities a bit more
ideally if I was smart, I'd be doing this where the music_db is actually
following the repository pattern. but I'm not so here we go.

not too happy with naming atm, but it'll do. main is now cleaner which
is nice.